+++ packages/storybook/examples/Tooltip/BasicTooltip.js
@@ -1,10 +1,24 @@
 import React from 'react';
 
-import Tooltip from '@ichef/gypcrete/src/Tooltip';
+import { PureTooltip as Tooltip } from '@ichef/gypcrete/src/Tooltip';
 
 function BasicTooltipExample() {
     return (
-        <Tooltip>tooltip</Tooltip>
+        <div>
+            <Tooltip>tooltip</Tooltip>
+
+            <div style={{ height: 30 }} />
+
+            <Tooltip placement="bottom">
+                placed at bottom of target
+            </Tooltip>
+
+            <div style={{ height: 30 }} />
+
+            <Tooltip arrowStyle={{ left: '12px' }}>
+                custom arrow style
+            </Tooltip>
+        </div>
     );
 }